94 separate changes:
Brought back in the changes made in jconfigure 1.311 (the ones that
were reverted out in jconfigure 1.312).

Got rid of most of the remaining runs of "echo" statements in
favor of here-documents ("cat <<- EOF" ... "EOF").   This simplifies the
quoting syntax and removes clutter.

Various cleanups of the output format.

Expanded English contractions (I'm => I am, and so forth) in order to
avoid potential trouble with quoting the apostrophe character.