Fixed a bug where remset entries were being generated during GC (by the stack scanning code), but were not being flushed into the remset pool correctly.
THis is a long standing problem.  I have now introduced an invariant that
remset entires are never generated during MMTk code.   If the host JVM
generates entires during the course of the GC, then it is the responsibility
of the host JVM to flush those entries to the remset pools.