Fix to baseline GC maps.  Kudos to Daniel for this.
The problem is that when jsr maps are merged with the parent normal map, the
use of a slot as a code reference should dominate: the entry should not show
up in both code and object reference maps.  This fix excludes an entry from
the mereged reference map if it is a code reference.