patch 1551835 : Support for JNI_OnLoad.
This patch changes the definition of the "syscall magic" so that the
first argument to the magic syscall is now the address of the function
(or in the case of POWEROPEN_API, the function description) that
should be invoked.  VM_SysCall is now a wrapper class that loads the
proper target address from the bootrecord and invokes the real magic.

Once the syscall magic is changed, it can then be used more generally to
invoke dynamically loaded C functions (in particular JNI_OnLoad) functions
as syscalls.