patch 1565581: Fix regalloc in spill code pass.
The issue is that we always assign non-volatile GPRs in reverse order
(to enable use of store/load multiple on PPC).  The scratch register code
was incorrectly iterating them in forward (non-reverse) order, which in turn
resulted in improper prologue/epilogue code being generated.  This patch
changes the scratch register code to iterate the registers in reverse.