fix bug in recomputeNormalOut.
If a branch instruction or fall through has been introduced whose target is an ExceptionHandlerBasicBlock,then we must set the bit on the target exception handler basic block that indicates it has a normal in edge.  Failure to do this results in improper computation of the normal out edges of the source block.