RVM-860 : Improper instruction format usage in PPC BURS rules
In the rules that take combinations of NEG(ADD(MULL(r, r), r)) and
produce a FNMADD instruction, the result of the NEG instruction
was being improperly accessed as if NEG was a Binary instead of
a Unary.  Wouldn't have actually mattered in terms of what
code was being generated, since the Result operand is at the same
offset in both Unary and Binary instruction forms, but the code
was incorrect.