RVM-875 : Fix the bug that causes intermittent ClassFormatErrors when running Xalan.
The bug is caused by a missing check in Atom.findOrCreate(byte[],boolean,String): After entering the synchronized block, we need to check if a matching Atom was already created.

Without the check, the following can happen: T1 checks for a matching Atom, finds none, and enters the synchronized block. Meanwhile, T2 checks for a matching Atom, finds none, and tries to enter the synchronized block but is blocked. T1 creates a new Atom in the synchronized block and returns it to its caller, allowing T2 to enter the synchronized block. T2 enters the synchronized block and creates a new Atom which is then returned to its caller. At that point, two distinct Atoms have been created for the same data which is not allowed.