RVM-750 : Implement Thread.setPriority (patch by Carl Ritson).
Carl Ritson's implementation uses pthread_setschedparam() to set the thread priority via the POSIX threads API. Another important point is that the implementation of  sysGetThreadPriority() always returns the OS priority of the thread instead of  assuming that the last call to sysSetThreadPriority() suceeded.

On Linux, it is necessary to use setpriority() to manipulate the nice level of the thread instead, at least for threads running in the default SCHED_OTHER policy. This requires obtaining the thread ID (thread's process id) using the gettid() syscall. One side effect of this method is that it is not possible to achieve priorities higher than NORM_PRIORITY on Linux without root access (or changing the scheduling policy).

Note that this implementation is different from the one in MRP.