Change debugging with DBG_SELECTIVE so that methods are printed when the original method of the current GenerationContext matches the method to print.
This ensures complete printouts for the method that is supposed to be debugged. Previously, printouts were deactivated as soon as a method was inlined because the generation of the inlined code via BC2IR reset the DBG_SELECTED variable.

Note that this change also means that BC2IR will not print out code for a method foo that is inlined into a method bar even if foo is contained in the sets of methods to print.