Fix JavaDoc errors in AOSOptions by changing the templates to not use characters that need escaping.
Note that it is not desirable to just escape the characters in the templates because the option description is also used to generate the command line help (which does not require escaping).

AFAIK it is not possible to modify GenerateFromTemplate to do the escaping during the generation of the option files without larger changes. GenerateFromTemplate does not seem to be able to distinguish the option descriptions from other text.