RVM-953 : Remove thread limit from Immix collectors.
It seems that the thread limit for the Immix collectors was only needed to restrict the size of the spillMarkHistograms in org.mmtk.policy.Defrag. The histograms are not needed at bootimage build time. This commit removes the Immix thread limit and allocates the histograms at runtime, immediately before collection is enabled.