RVM-1105 : Remove libvm (original changes by Ian Rogers and Da Feng for MRP).
This commit removes the libvm.c files and distributes the functionality into other parts of the bootloader. The mapping of image files and the creation of the VM is moved into main.c . Code to start threads and to handle signals is moved to sysThread.c and the new sys* files. The commit also changes the setup of threads so that only one signal handling stack is allocated.

The MRP commits that have been used for this commit are:
- https://github.com/codehaus/mrp/commit/85a2c78266ab3121cb1b49f4fd3a8369cf685563 (libvm removal, part 1)
- https://github.com/codehaus/mrp/commit/0d8544113769abcfb54e116339b5311ad79e745e (libvm removal, part 2)
- https://github.com/codehaus/mrp/commit/6784db17395a1e3c1410c2ab5bb5eb8c69f81ecd (PPC build fixes)
- https://github.com/codehaus/mrp/commit/50efc8345defb218a6e365788675d83a4046f070 (fixes for sysSignal_ppc.c)
- https://github.com/codehaus/mrp/commit/155f1c20571b56e8598ba82d4fdb9ee4cc83867c (PPC signal macro changes)
- https://github.com/codehaus/mrp/commit/e2d03c9eaaab2e112107e2c545102799d1fe3e72 (fixes for mapImageFile and main)
- https://github.com/codehaus/mrp/commit/ba3c07869445954b5960f7001e9bb1042ac1d7ee (Da Feng's bugfix for register encoding in IA32 64-bit signal handler)
- https://github.com/codehaus/mrp/commit/071eb5df19f0d33ce99a7cf9266ddf7a27af8bcf (printf format fixes, Mac OS IA32 dumpContext fix). That MRP commit also changes the data type for image map file sizes from "unsigned" to "long" in createVM(..). This fixes unaligned stores in mapImageFile which caused undefined behaviour.

The commit also has some fixes for problems in sysSignal_ppc.c that are not fixed in MRP:
- readContextInformation(..) in has been changed to use the thread register to set threadPtr
- readContextTrapCode(..) has been changed to not use the UNUSED macro for the context parameter because it's actually being used
- the impementation for setupDeliverHardwareException(..) was replaced with code to call RuntimeEntrypoints.deliverHardwareException(..) instead of setupDumpStackAndDie(..)

There are also some minor differences with respect to the original MRP version of the code.  For example, in our case sysInitialize() is called after populating the boot record because we want to use checkMalloc(..) in sysInitialize() which requires that the boot record is already set up. Moreover, the naming of the architecture specific sys*.c files has been changed to use the ${target.arch} variable in order to be more consistent with the naming of our Java packages.

Almost all of the rest of the code is directly adapted from MRP.