Fail the boot image build immediately when an unexpected OptimizingCompilerException occurs.
This ensures that broken builds fail fast. Failing fast ensures that the
build log doesn't get filled up with redundant (debug) information.

Not needlessly dumping information is very important. For example, the
bug that was fixed in the last commit lead to log files that filled up
the complete hard disk on a regression machine (300+ GB).