RVM-153 : Convert org.mmtk.utility.heap.PageResource#ZERO_ON_RELEASE to a build time constant and set it to true in the stress test run.
This is implemented by controlling the zeroing of pages with a MMTk property. Prior to this commit, the MMTk property file could only be set by passing the value as argument to Ant (e.g. -Dmmtk.config=sideMarkBit). The changes in this commit allow Jikes RVM config files (e.g. for prototype) to set the mmtk.config property. This allows the gcstress configuration to use a custom MMTk property file that ensures that pages are zeroed on release.