Add an Ant task that will generate a summary file about the boot image sizes when running test-runs.
The summary file is a UTF-8 encoded .csv file intended for use with the Jenkins plot plugin to graphically track image sizes. The data is generated by computing the total image size for each boot image in the images directory (e.g. dist/ by default). The .csv file is encoded in UTF-8 because that seems to be the encoding that Jenkins expects.

The first line in the csv file contains the image directory name, the second line contains the total image size in bytes. Example file contents:
prototype-opt_x86_64-linux,prototype_x86_64-linux
39762938,23999512