RVM-1099 : Change instantiation of types during boot image building to use multiple threads by default.
The order of instantiation is now explicitly managed in order to avoid
deadlocks. Tested on ia32 32-bit dual core and on ppc64 with 32 threads.

In the new code, java.lang.Object is instantiated before all other types
because it is the root of the hierarchy and it doesn't make sense to
defer its instantiation.