Restructure treatment of thread stack sizes.
Previously, the RVMThread class was responsible for mapping the stack size passed in from java.lang.Thread to the internal stack size. This has been moved to VMCommonLibrarySupport because the same problem needs to be solved in OpenJDK. Keeping the code in the constructor of RVMThread makes it hard to read. Moreover, all other callers of the RVMThread constructors are internal Jikes RVM classes which can be expected to pass correct stack sizes that don't need adjustment.

This commit also adds an assertion for invalid stack sizes in MemoryManager.newStack(..).