RVM-1131 : Change getMethods() in the Jikes RVM implementation of java.lang.Class to return multiple methods for interfaces.
Quote from the Java API: "If the class declares multiple public member
methods with the same parameter types, they are all included in the
returned array.". According to the StarInheritance test in the openjdk
test run, this means that all non-overriden definitions that reach a
class from its superclasses or interfaces must be present in the array.
This means that methods with the same name-signature combination can be
present multiple times for interfaces and abstract classes.

This fixes a failure in the StarInheritance test. Note that the test
still doesn't pass in the openjdk test run: The openjdk test run
uses checksums to check the output and the order of methods in our
implementation is different. The API spec doesn't impose a certain
ordering.