RVM-1092 : Use replaceOperandWithSpillLocation(..) in SpillCode for YIELDPOINT_OSR when running on IA32.
For IA32, the operand must be replaced with a spill location in YIELDPOINT_OSR. Not doing so can cause YIELDPOINT_OSR instructions to have operands which no longer have any definitions which is obviously incorrect. These kinds of failures can be found with enabled paranoid IR verification.

For PPC, it is *probably* correct to keep the behaviour of ignoring YIELDPOINT_OSR. The history of the changes on PPC are as follows:
- 673d2cd1fcc2b3f20704a26775b1f08595048496 ("fix for scratch registers with OSR") changed the PPC StackManager to return false for needScratch(..) for YIELDPOINT_OSR
- a2981a49741d10d4e35feee74c5a4ef861333473 ("ignore YIELDPOINT_OSR during spill code generation") changed the spill code generation to ignore YIELDPOINT_OSR for PPC *and* IA32
- ac8bcb373195fe1e21d6e1d775de808012900c70 ("patch 1354655 : more verifier found fixes (bug fixes to opt compiler to remove places where invalid IR was being created)") changed both IA32 *and* PPC to handle YIELDPOINTS_OSR instead of ignoring it in an attempt to fix problemes revealed by IR verification
- 68bafbda1652a8ae398e089f9425c9c13eaf504e ("back out previous change; Robin reports the assertion triggers on Linux/PPC") reversed ac8bcb373195fe1e21d6e1d775de808012900c70 to get back to the behaviour of a2981a49741d10d4e35feee74c5a4ef861333473 because the new behaviour triggered the assertion that was supposed to be unreachable. The assertion was triggered due to needScratch(..) returning false for YIELDPOINT_OSR

If there are still failures for OSR related to spill code after this, somebody will have to take a closer look at the related code.