RVM-1092: Change Simple optimisation phase so that it removes all statements in a basic block that follow an unconditional trap.
This fixes a paranoid IR verification failure in the method headMap(java.lang.Object, boolean) from TreeMap. A must_implement_interface check was converted to an unconditional trap which eliminated the definition of a temporary. This led to subsequent instructions (which aren't reachable) having a use before def for that temporary.