RVM-549 : Extend IcedTea build file so that it can be used in place of the OpenJDK build file.
Note that it's not possible to actually build with IcedTea 6: the build will fail in the "compile" target. The current OpenJDK port replaces some JDK classes with classes from the libraryInterface directory. Those replacement classes are based on the OpenJDK zip which contains an older version of the JDK. This leads to incompatibilities that cause compile errors.