Change JNIEnvironment to ensure that all new JNIRefsArrays are allocated into the non-moving space.
This fixes an assertion failure in ScanThread that I saw while working on the OpenJDK port. The method assertImmovableInCurrentCollection() in ScanThread checks that the JNIRefs array is non-movable. This was only the case for first JNIRefs array in each JNIEnvironment. When the JNIRefs arrays had to be grown and new address arrays were allocated, those new arrays were wrongly allocated into the default collector spaces.