RVM-1063 : Move responsibility for determining type of values from GenericStackManager to architecture-specific StackManagers.
Defining architecture-independent value types unnecessarily constrains the architecture-specific stack managers. Moreover, the types of values that are interesting are architecture-dependent. For example, the x64 StackManager needs to distinguish between 32-bit values and 64-bit values in order to make sure that non-quad instructions are not falsely promoted to quad instructions.

For now, the code can just rely on the types provided by Register which provide all the information that the StackManagers for the currently supported architectures require.