RVM-1062 : Clear MSBs when emitting ireturn().
The RET instruction is 64-bit on x64 so the compiler must make sure that the upper 32 bits of the register are cleared when returning ints.

This fixes a failure on x64 development hsqldb. The returned int was used as an array index. This caused a failure because the upper 32 bits weren't cleared.

The first lines of the stack trace were:
Failing instruction offset: 0x0000000000000094 in method get with descriptor (Ljava/lang/Object;I)I declared by class with descriptor Lorg/hsqldb/lib/IntValueHashMap;
Couldn't find a method for given instruction offset