Attempt to make sure that jburg closes each output file or flushes stdout.
In regression tests, the build sometimes fails at the compile stage because one of the output files isn't correctly written. That seems to very rare on IA32 (I only saw it once) but it happens on the PPC64 tests every few test runs.

From looking at the code, it seems to me that most files are closed properly. The initial outfp file stream is neither closed nur flushed, so that's the first thing to change. I'm not sure if this fixes the problem but it's worth a try.