Fix "bad GC map" problems in null check tests in gcstress builds that compile the test methods with the baseline compiler.
The issue was that the stack state didn't always match the state that the ReferenceMaps expect. In particular, most cases for null checks do the access in an atomic way so the stack isn't changed when a NPE occurs. This wasn't the case for all field accesses.

The solution is to not change the stack state until it is ensured that no NPE will happen.