Remove duplication from verbose outputs in DefaultInlineOracle.
The new implementation causes the messages to be always built, even if the output is never printed. For the new method reportInitialProfileState(..), this would lead to a NPE during boot image building for prototype-opt because the DCG isn't available then. That's why reportInitialProfileState(..) has an additional verbosity check. This will be fixed in the next commit.