RVM-1156 : Fix bad GC map for long parameters in x64 dynamic bridge frames.
The problem occurred for methods that have a long parameter followed by a reference parameter. It was caused by inconsistent handling of the size of long parameters. Long parameters take 1 register word on x64 instead of 2 as on IA32 32-bit. This applies both for the bridge register index and the bridge register location.

The bridge register index was handled correctly for x64 (i.e. it was increased by 1). Therefore, the iterator correctly determined that the reference parameter was orginally in the parameter GPR T1 and later saved to the stack. However, the iterator modified the bridge register location by 2 stack words instead of 1. This caused the baseline GC map iterator to treat stack slots as references that were in fact not references. The iterator read the object parameter from the XMM save area instead:

Stack for x64 dynamic bridge frame for affected method
<- CMID ->
<- EDI  ->
<- EBX  ->
<- T0   -> <-- first parameter  (long)
<- T1   -> <-- second parameter (reference)
<- XMM3 -> <-- stack location that was read instead of T1's location