RVM-1156 : Add incomplete unit test for IA32 ArchBridgeDataExtractor.
The unit test only tests the bug that causes RVM-1156. The behaviour of the affected code differs for builds with 32-bit addressing and builds with 64-bit addressing. That's why this commit introduces new annotations for the test runner to distinguish the addressing modes.

It's also necessary to make the IA32 ArchBridgeDataExtractor non-final to be able to test the code.