RVM-334 : Send SIGQUIT before terminating process in timedrun script.
Sending SIGQUIT will signal to the VM that a thread dump should occur.

The timedrun script currently uses Perl's system() function to run the tests. This functions block SIGQUIT. Therefore the bootloader needs to unblock SIGQUIT.

This commit also adds a test that can be run to determine if the timedrun script works correctly.