RVM-1062 / RVM-1063 : Fix bugs in x64 argument passing for syscalls.
Previously, longs and doubles were incorrectly taking two stack slots for syscalls in the baseline compiler. Additionally, in the baseline compiler the argument pointer wasn't adjusted for arguments passed on the stack. This lead to the pointer pointing to the wrong arguments when subsequent arguments were passed in registers. For example, when all GPRs were full and more integer arguments were passed on the stack, the argument pointer was wrong for subsequent floating point arguments (which were passed in FPRs).

The argument passing for x64 syscalls in the optimizing compiler has been rewritten to fix some bugs and to make it slighly easier to understand.