Add library interface class for java.lang.System to fix infinite loop.
The inifite loop was caused by a missing implementation of System.arraycopy(..). The previous implementation of System.arrayCopy was called, but it was native, so method compilation had to take place, which caused use of StringBuilder, which in turn uses System.arraycopy(..). After removing the string concatenation from method compilation, there was still a StringBuilder usage in name mangling (via normal +). In order to fix this, use a non-native method for arraycopy(..) just as we do for GNU Classpath.