RVM-549 : Change classloading so that loading of replacement classes is only attempted for classes that actually have a replacement
The previous approach was inefficient because loading of replacement classes was attempted for each class loaded with the bootstrap classloader.

The set of replacement classes is determined at build time with an annotation processor. The set is passed to the bootstrap classloader through a generated class, similar to SysCallImpl. This ensures that the set of replacement classes won't be modified after build time (which was never supported anyway) and makes sure that the set is current.

Note that the annotation processor doesn't validate if the replacement class is semantically viable. This is a deliberate design choice. The annotation processor doesn't have access to the same information that the VM has at classloading time. In particular, it might not be able to access all implementation details of classes.