RVM-549 : Add implementations for java.lang.Object and java.lang.Class
For java.lang.Object, we can't use the replacement class mechanism. An attempt to do so causes bootstrap problems because the classloading will attempt to do annotation processing which requires that java.lang.Object is resolved.

For java.lang.Class, the boot image writer needs to be able to set up the RVMType when copying the objects. This is achieved by naming the RVMType field in java.lang.Class as "type" as in GNU Classpath. This leads to re-using the logic from GNU Classpath in the boot image writer.