RVM-549 : Add code to the boot image writer to copy JDK 8 HashMaps to JDK 6 HashMaps
The internal structure of the HashMap class was changed in OpenJDK 8. Before this commit, the table field from the HashMap class was nulled out because of differing types. This broke the HashMap instances.

HashMaps in JDK 8 use an array of HashMap$Node and HashMaps in JDK 6 use an array of HashMap$Entry. The classes are similar enough that we can just read out fields for HashMap$Entry from HashMap$Node instances. However, the hash values and the indexes in the array change for JDK 6 because the hash function used in HashMap also differs between JDK 8 and JDK 6. Consequently, the boot image writer has to rebuild the table array for JDK 6.