RVM-1177 : Correct handling of assertion arguments in bootloader
Assertions arguments are passed to Java by the bootloader and thus must be explicitly mentioned in the list of VM directives in the bootloader.

This fixes invocations like "./rvm -ea -Xmx100M SomeClass" which previously errored out with a message that claimed "-Xmx100M" was not a recognized Jikes RVM command line argument.