RVM-549 : Keep libraries for OpenJDK in their normal OpenJDK locations
Previously, libraries from OpenJDK where copied to the base directory of the distribution (dist). This was inconsistent with the normal approach in OpenJDK (load from lib/amd64 on x64). Additionally, it caused loading libraries for awt to fail.

This commit changes the organization of libraries to keep them at their usual place for OpenJDK. This requires changes to the loading of libjava for OpenJDK 6 because libjava needs to be loaded before loading via System.load(..) / System.loadLibrary(..) works. The loading of libjava is done via DynamicLibrary which needs an absolute path to find libjava in a subdirectory (i.e. lib/$arch/ in the distribution directory). Therefore, the absolute path to the distribution directory is now passed in the rvm start script.

Additionally, this commit replaces "amd64" with a new variable for the OpenJDK architecture.